InjectionModel: Removed the limit on the number of particles per parcel
A lower limit of one on the number of particles represented by a single
parcel has been removed from the injection models. It may be appropriate
to simulate the statistical behaviour of a particulate flow with more
lagrangian elements than physical particles. A unity lower limit does
not permit this.

The limit was, in some situations, also causing the large-diameter end
of an injected distribution to be clipped.

This resolves bug report https://bugs.openfoam.org/view.php?id=2837