phaseTransferModels: Added deposition model
This model transfers a dispersed droplet phase to a film phase at a rate
relative to its intersection with a third phase. The third phase is
termed the "surface". It can be enabled in constant/phaseProperties as
follows:

    phaseTransfer
    (
        (droplets and film)
        {
            type            deposition;
            droplet         droplets;
            surface         solid;
            efficiency      0.1;
        }
    );

The efficiency is an empirical factor which represents a reduction in
collisions as a result of droplets flowing around the surface phase and
not coalescing on impact.

This work was supported by Georg Skillas and Zhen Li, at Evonik