particleForces: New scaled force, and optional alternate syntax
In addition to the current syntax, particle forces can now also be
specified with a named dictionary and a type entry. This allows multiple
forces of the same type to be used. For example:

    particleForces
    {
        // Existing syntax. The model typename, either on it's own
        // or heading a dictionary of model parameters.

        gravity;

        ErgunWenYuDrag
        {
            alphac alpha.air;
        }

        // New syntax. An arbitrary name, followed by a dictionary
        // containing a type entry, and any model parameters.

        gravity2
        {
            type gravity;
        }

        drag2
        {
            type ErgunWenYuDrag;
            alphac alpha.air;
        }
    }

A scaled force has also been added. This is a wrapper around another
particle force model, and applies a scaling factor to the force value.
It can be used to tune the models, or to blend two or more models.

    particleForces
    {
        // Tuning the simulation by reducing the pressure gradient force

        scaled1
        {
            type scaled;
            factor 0.94;
            forceType pressureGradient;
            U U;
        }

        // Blending two drag models

        scaled2
        {
            type scaled;
            factor 0.2;
            forceType sphereDrag;
        }

        scaled3
        {
            type scaled;
            factor 0.8;
            forceType ErgunWenYuDrag;
            alphac alpha.air;
        }
    }

This work was supported by Moritz Hoefert, at Evonik