localEulerDdtScheme: Added fvcDdtPhiCoeff calculated from the local time-step
For cases where the time-step varies in space ddtCorr is not needed with LTS but
in regions where the time-step is uniform it may be required.  This change
ensures that the localEuler scheme operates equivalently to Euler in regions and
for cases where the time-step is uniform.

Addresses report https://bugs.openfoam.org/view.php?id=3020